# specs/001-job-portal-basic/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Student Job Market API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /api/jobs:
    get:
      summary: List approved jobs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobListItem"
    post:
      summary: Create a new job (requires login)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /api/jobs/{id}:
    get:
      summary: Get job detail
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found
...
components:
  schemas:
    JobListItem:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        field: { type: string }
        createdAt: { type: string, format: date-time }
    Job:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        salary: { type: string }
        startDate: { type: string, format: date }
        employmentType: { type: string, enum: [full-time, part-time] }
        field: { type: string }
        email: { type: string, format: email }
        status: { type: string, enum: [pending, approved] }
        createdAt: { type: string, format: date-time }
    JobCreate:
      type: object
      required: [title, company, location, employmentType, field, email]
      properties:
        title: { type: string }
        company: { type: string }
        location: { type: string }
        salary: { type: string }
        startDate: { type: string, format: date }
        employmentType: { type: string, enum: [full-time, part-time] }
        field: { type: string }
        email: { type: string, format: email }
...